added api files
api.py is the actual code I've done so far with Flask. The zip file contains the entire api folder. It has a hidden .env folder and the python virtual environment inside. I just thought it might be useful to unzip the api folder into the project and make sure it works on everyone's machine.